whoops

+++ ProjectTomato/src/lib.rs
@@ -1,6 +1,6 @@
 use jni::JNIEnv;
 
 #[no_mangle]
-pub extern "C" fn Java_org_hermitsocialclub_tomato_LibTomato_splat(_env: JNIEnv) -> i64 {
+pub extern "C" fn Java_org_hermitsocialclub_tomato_LibTomato_splat(_env: JNIEnv) -> i32 {
     69
 }
